refactor: 상속관계 변경 - ProductInventory를 서브클래스로 변경하였습니다.
